feat(scion-rcp): extract scion microfrontend API interfaces
 * Faciliates decorating these services.
