+++ ch.sbb.scion.rcp.microfrontend.app.demo/src/ch/sbb/scion/rcp/microfrontend/app/demo/parts/DevToolsPart.java
@@ -8,8 +8,8 @@ import org.eclipse.swt.SWT;
 import org.eclipse.swt.widgets.Composite;
 
 import ch.sbb.scion.rcp.microfrontend.OutletRouter;
-import ch.sbb.scion.rcp.microfrontend.OutletRouter.NavigationOptions;
 import ch.sbb.scion.rcp.microfrontend.RouterOutlet;
+import ch.sbb.scion.rcp.microfrontend.model.NavigationOptions;
 import ch.sbb.scion.rcp.microfrontend.model.Qualifier;
 
 public class DevToolsPart {
