+++ ch.sbb.scion.rcp.microfrontend.app.demo/src/ch/sbb/scion/rcp/microfrontend/app/demo/parts/IntentClientPart.java
@@ -54,8 +54,8 @@ import org.eclipse.swt.widgets.Text;
 
 import com.google.gson.Gson;
 
-import ch.sbb.scion.rcp.microfrontend.IntentClient;
 import ch.sbb.scion.rcp.microfrontend.MessageClient;
+import ch.sbb.scion.rcp.microfrontend.IntentClient;
 import ch.sbb.scion.rcp.microfrontend.app.demo.parts.model.IntentMessageModel;
 import ch.sbb.scion.rcp.microfrontend.app.demo.parts.model.IntentSelectorModel;
 import ch.sbb.scion.rcp.microfrontend.model.Intent;
