+++ ch.sbb.scion.rcp.microfrontend.e3.app.demo/src/ch/sbb/scion/rcp/microfrontend/e3/app/demo/view/OutletRouterPart.java
@@ -13,7 +13,7 @@ import org.eclipse.swt.SWT;
 import org.eclipse.swt.widgets.Composite;
 
 import ch.sbb.scion.rcp.microfrontend.OutletRouter;
-import ch.sbb.scion.rcp.microfrontend.OutletRouter.NavigationOptions;
+import ch.sbb.scion.rcp.microfrontend.model.NavigationOptions;
 
 public class OutletRouterPart {
