+++ ch.sbb.scion.rcp.microfrontend/META-INF/MANIFEST.MF
@@ -18,14 +18,14 @@ Export-Package: ch.sbb.scion.rcp.microfrontend,
  ch.sbb.scion.rcp.microfrontend.interceptor,
  ch.sbb.scion.rcp.microfrontend.model,
  ch.sbb.scion.rcp.microfrontend.subscriber
-Service-Component: OSGI-INF/ch.sbb.scion.rcp.microfrontend.IntentClient.xml,
- OSGI-INF/ch.sbb.scion.rcp.microfrontend.ManifestService.xml,
- OSGI-INF/ch.sbb.scion.rcp.microfrontend.MessageClient.xml,
- OSGI-INF/ch.sbb.scion.rcp.microfrontend.MicrofrontendPlatform.xml,
- OSGI-INF/ch.sbb.scion.rcp.microfrontend.OutletRouter.xml,
- OSGI-INF/ch.sbb.scion.rcp.microfrontend.host.IntentInterceptorInstaller.xml,
+Service-Component: OSGI-INF/ch.sbb.scion.rcp.microfrontend.host.IntentInterceptorInstaller.xml,
  OSGI-INF/ch.sbb.scion.rcp.microfrontend.host.MessageInterceptorInstaller.xml,
  OSGI-INF/ch.sbb.scion.rcp.microfrontend.host.MicrofrontendPlatformRcpHost.xml,
+ OSGI-INF/ch.sbb.scion.rcp.microfrontend.internal.IntentClientImpl.xml,
+ OSGI-INF/ch.sbb.scion.rcp.microfrontend.internal.ManifestServiceImpl.xml,
+ OSGI-INF/ch.sbb.scion.rcp.microfrontend.internal.MessageClientImpl.xml,
+ OSGI-INF/ch.sbb.scion.rcp.microfrontend.internal.MicrofrontendPlatformImpl.xml,
+ OSGI-INF/ch.sbb.scion.rcp.microfrontend.internal.OutletRouterImpl.xml,
  OSGI-INF/ch.sbb.scion.rcp.microfrontend.keyboard.KeyboardEventMapper.xml
 Import-Package: javax.annotation,
  javax.inject;resolution:=optional,
