+++ ch.sbb.scion.rcp.microfrontend/OSGI-INF/ch.sbb.scion.rcp.microfrontend.internal.ManifestServiceImpl.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.3.0" name="ch.sbb.scion.rcp.microfrontend.ManifestService">
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.3.0" name="ch.sbb.scion.rcp.microfrontend.internal.ManifestServiceImpl">
    <service>
       <provide interface="ch.sbb.scion.rcp.microfrontend.ManifestService"/>
    </service>
    <reference cardinality="1..1" field="microfrontendPlatformRcpHost" interface="ch.sbb.scion.rcp.microfrontend.host.MicrofrontendPlatformRcpHost" name="microfrontendPlatformRcpHost"/>
-   <implementation class="ch.sbb.scion.rcp.microfrontend.ManifestService"/>
+   <implementation class="ch.sbb.scion.rcp.microfrontend.internal.ManifestServiceImpl"/>
 </scr:component>
