+++ ch.sbb.scion.rcp.microfrontend/OSGI-INF/ch.sbb.scion.rcp.microfrontend.internal.MessageClientImpl.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.3.0" name="ch.sbb.scion.rcp.microfrontend.MessageClient">
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.3.0" name="ch.sbb.scion.rcp.microfrontend.internal.MessageClientImpl">
    <service>
       <provide interface="ch.sbb.scion.rcp.microfrontend.MessageClient"/>
    </service>
    <reference cardinality="1..1" field="microfrontendPlatformRcpHost" interface="ch.sbb.scion.rcp.microfrontend.host.MicrofrontendPlatformRcpHost" name="microfrontendPlatformRcpHost"/>
-   <implementation class="ch.sbb.scion.rcp.microfrontend.MessageClient"/>
+   <implementation class="ch.sbb.scion.rcp.microfrontend.internal.MessageClientImpl"/>
 </scr:component>
