+++ ch.sbb.scion.rcp.microfrontend/OSGI-INF/ch.sbb.scion.rcp.microfrontend.internal.OutletRouterImpl.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.3.0" name="ch.sbb.scion.rcp.microfrontend.OutletRouter">
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.3.0" name="ch.sbb.scion.rcp.microfrontend.internal.OutletRouterImpl">
    <service>
       <provide interface="ch.sbb.scion.rcp.microfrontend.OutletRouter"/>
    </service>
    <reference cardinality="1..1" field="microfrontendPlatformRcpHost" interface="ch.sbb.scion.rcp.microfrontend.host.MicrofrontendPlatformRcpHost" name="microfrontendPlatformRcpHost"/>
-   <implementation class="ch.sbb.scion.rcp.microfrontend.OutletRouter"/>
+   <implementation class="ch.sbb.scion.rcp.microfrontend.internal.OutletRouterImpl"/>
 </scr:component>
