+++ ch.sbb.scion.rcp.workbench/src/ch/sbb/scion/rcp/workbench/popup/Popup.java
@@ -17,8 +17,8 @@ import ch.sbb.scion.rcp.microfrontend.model.TopicMessage;
 import ch.sbb.scion.rcp.microfrontend.subscriber.ISubscriber;
 import ch.sbb.scion.rcp.microfrontend.subscriber.ISubscription;
 import ch.sbb.scion.rcp.workbench.IWorkbenchPopup;
-import ch.sbb.scion.rcp.workbench.WorkbenchPopupSize;
 import ch.sbb.scion.rcp.workbench.WorkbenchPopupOrigin;
+import ch.sbb.scion.rcp.workbench.WorkbenchPopupSize;
 import ch.sbb.scion.rcp.workbench.internal.ContextInjectors;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
